Single-Garment-Mood-image/comfy-alle-custom-nodes/comfy-alle-custom-nodes/requirements.txt
python-dotenv==1.0.1
pillow==10.3.0
requests==2.31.0